chore: move to one package structure
