+++ demo/src/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import CodeExample from './components/CodeExample.vue'
 import { snippetz } from '@scalar/snippetz'
-import { type ClientId, type TargetId } from '@scalar/snippetz-core'
+import { type ClientId, type TargetId } from '@scalar/snippetz/core'
 import { ref } from 'vue'
 
 const selectedTarget = ref<TargetId>('node')
