--- packages/snippetz-core/CHANGELOG.md
@@ -1,32 +1,0 @@
-# @scalar/snippetz-core
-
-## 0.1.4
-
-### Patch Changes
-
-- 893d1af: add ofetch
-
-## 0.1.3
-
-### Patch Changes
-
-- e352e5a: feat: add js fetch
-
-## 0.1.2
-
-### Patch Changes
-
-- fbf21d2: feat: add snippetz-plugin-node-fetch
-
-## 0.1.1
-
-### Patch Changes
-
-- 31b9be4: fix: JSON.stringify for undici body
-- 9d0c097: chore: rename snippetz-plugin-undici to snippetz-plugin-node-undici
-
-## 0.1.0
-
-### Minor Changes
-
-- 92b5028: init :)
