--- packages/snippetz-plugin-js-ofetch/CHANGELOG.md
@@ -1,21 +1,0 @@
-# @scalar/snippetz-plugin-js-ofetch
-
-## 0.1.1
-
-### Patch Changes
-
-- 893d1af: add ofetch
-- 893d1af: add ofetch
-- Updated dependencies [893d1af]
-  - @scalar/snippetz-core@0.1.4
-
-## 0.1.0
-
-### Minor Changes
-
-- 0bef889: init
-
-### Patch Changes
-
-- Updated dependencies [e352e5a]
-  - @scalar/snippetz-core@0.1.3
