--- packages/snippetz-plugin-node-fetch/tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src"],
-  "exclude": ["dist", "test", "vite.config.ts", "**/*.test.ts"],
-  "compilerOptions": {
-    "declaration": true,
-    "declarationMap": true,
-    "emitDeclarationOnly": true,
-    "outDir": "dist/",
-    "allowSyntheticDefaultImports": true
-  }
-}
