--- packages/snippetz-plugin-node-fetch/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {},
-  "include": ["src", "rollup.config.ts"]
-}
