--- packages/snippetz-plugin-node-ofetch/CHANGELOG.md
@@ -1,28 +1,0 @@
-# @scalar/snippetz-plugin-node-fetch
-
-## 0.1.1
-
-### Patch Changes
-
-- 893d1af: add ofetch
-- 893d1af: add ofetch
-- Updated dependencies [893d1af]
-  - @scalar/snippetz-core@0.1.4
-
-## 0.1.0
-
-### Patch Changes
-
-- Updated dependencies [e352e5a]
-  - @scalar/snippetz-core@0.1.3
-
-## 0.1.0
-
-### Minor Changes
-
-- 6d1dd21: init :)
-
-### Patch Changes
-
-- Updated dependencies [fbf21d2]
-  - @scalar/snippetz-core@0.1.2
