--- packages/snippetz-plugin-node-undici/CHANGELOG.md
@@ -1,59 +1,0 @@
-# @scalar/snippetz-plugin-node-undici
-
-## 0.1.6
-
-### Patch Changes
-
-- Updated dependencies [893d1af]
-  - @scalar/snippetz-core@0.1.4
-
-## 0.1.5
-
-### Patch Changes
-
-- Updated dependencies [e352e5a]
-  - @scalar/snippetz-core@0.1.3
-
-## 0.1.4
-
-### Patch Changes
-
-- Updated dependencies [fbf21d2]
-  - @scalar/snippetz-core@0.1.2
-
-## 0.1.3
-
-### Patch Changes
-
-- 31b9be4: fix: JSON.stringify for undici body
-- 9d0c097: chore: rename snippetz-plugin-undici to snippetz-plugin-node-undici
-- Updated dependencies [31b9be4]
-- Updated dependencies [9d0c097]
-  - @scalar/snippetz-core@0.1.1
-
-## 0.1.2
-
-### Patch Changes
-
-- eca1c31: chore: add a basic plugin architecture
-- Updated dependencies [92b5028]
-  - @scalar/snippetz-core@0.1.0
-
-## 0.1.1
-
-### Patch Changes
-
-- 9d4fedd: chore: remove prettier, just use strings
-- Updated dependencies [9d4fedd]
-  - @scalar/snippetz@0.1.1
-
-## 0.1.0
-
-### Minor Changes
-
-- 64a09b0: init :)
-
-### Patch Changes
-
-- Updated dependencies [64a09b0]
-  - @scalar/snippetz@0.1.0
