+++ packages/snippetz/package.json
@@ -15,14 +15,7 @@
     "test:ui": "vitest --ui",
     "types:check": "tsc --noEmit --skipLibCheck"
   },
-  "dependencies": {
-    "@scalar/snippetz-core": "workspace:*",
-    "@scalar/snippetz-plugin-js-fetch": "workspace:*",
-    "@scalar/snippetz-plugin-js-ofetch": "workspace:^",
-    "@scalar/snippetz-plugin-node-fetch": "workspace:*",
-    "@scalar/snippetz-plugin-node-ofetch": "workspace:^",
-    "@scalar/snippetz-plugin-node-undici": "workspace:*"
-  },
+  "dependencies": {},
   "files": [
     "dist"
   ],
