+++ packages/snippetz/rollup.config.ts
@@ -1,0 +1,15 @@
+import {
+    addPackageFileExports,
+    createRollupConfig,
+} from '@scalar/build-tooling'
+
+const entries = ['src/index.ts']
+
+export default createRollupConfig({
+    typescript: true,
+    options: {
+        input: entries,
+    },
+})
+
+await addPackageFileExports({ entries })
