+++ packages/snippetz/src/core/utils/arrayToObject.ts
