+++ packages/snippetz/src/plugins/js/fetch/fetch.ts
@@ -3,9 +3,9 @@ import {
   type Request,
   arrayToObject,
   objectToString,
-} from '@scalar/snippetz-core'
+} from '../../../core'
 
-export function ofetch(request?: Partial<Request>): Source {
+export function fetch(request?: Partial<Request>): Source {
   // Defaults
   const normalizedRequest = {
     method: 'GET',
@@ -27,13 +27,7 @@ export function ofetch(request?: Partial<Request>): Source {
       ? arrayToObject(normalizedRequest.queryString)
       : undefined
   )
-
-  if (searchParams.size) {
-    options.query = {}
-    searchParams.forEach((value, key) => {
-      options.query[key] = value
-    })
-  }
+  const queryString = searchParams.size ? `?${searchParams.toString()}` : ''
 
   // Headers
   if (normalizedRequest.headers?.length) {
@@ -69,7 +63,9 @@ export function ofetch(request?: Partial<Request>): Source {
 
     // JSON
     if (normalizedRequest.postData.mimeType === 'application/json') {
-      options.body = JSON.parse(options.body)
+      options.body = `JSON.stringify(${objectToString(
+        JSON.parse(options.body)
+      )})`
     }
   }
 
@@ -79,12 +75,12 @@ export function ofetch(request?: Partial<Request>): Source {
     : ''
 
   // Code Template
-  const code = `ofetch('${normalizedRequest.url}'${jsonOptions})`
+  const code = `fetch('${normalizedRequest.url}${queryString}'${jsonOptions})`
 
   // Create an AST
   return {
-    target: 'node',
-    client: 'ofetch',
+    target: 'js',
+    client: 'fetch',
     code,
   }
 }
