+++ packages/snippetz/src/plugins/js/ofetch/index.ts
